+++ CHANGES.md
@@ -14,7 +14,7 @@ Core Changes:
 
 Language Improvements:
 
-- (markdown) improve bold/italic nesting () [Josh Goebel][]
+- (markdown) improve bold/italic nesting (#2382) [Josh Goebel][]
 - (fortran) Add Fortran 2018 keywords and coarray intrinsics (#2361) [Sam Miller][]
 - (delphi) highlight hexadecimal, octal, and binary numbers (#2370) [Robert Riebisch]()
 - enh(plaintext) added `text` and `txt` as alias (#2360) [Taufik Nurrohman][]